CBLT-42:  FAIL if setup.exe not made

+++ build/package-win.rb
@@ -75,10 +75,14 @@ INSTALL_PROJ  = "Sync_Gateway.ism"
 INSTALL_SRC   = "#{START_DIR}/windows/InstallShield_2014_Projects"
 INSTALL_OUT   = "#{INSTALL_SRC}/Sync_Gateway/PROJECT_ASSISTANT/SINGLE_EXE_IMAGE/DiskImages/DISK1"
 
-print "\nISCmdBld.exe -d ProductVersion=#{RELEASE}-#{BLDNUM},PATH_TO_JENKINS_WORKSPACE=ENV['WORKSPACE']  #{INSTALL_SRC}/#{INSTALL_PROJ}"
-        `ISCmdBld.exe -d ProductVersion=#{RELEASE}-#{BLDNUM},PATH_TO_JENKINS_WORKSPACE=ENV['WORKSPACE']  #{INSTALL_SRC}/#{INSTALL_PROJ}`
+print "\nISCmdBld.exe -d ProductVersion=#{RELEASE}-#{BLDNUM},PATH_TO_JENKINS_WORKSPACE=#ENV['WORKSPACE']  #{INSTALL_SRC}/#{INSTALL_PROJ}"
+        `ISCmdBld.exe -d ProductVersion=#{RELEASE}-#{BLDNUM},PATH_TO_JENKINS_WORKSPACE=#ENV['WORKSPACE']  #{INSTALL_SRC}/#{INSTALL_PROJ}`
 
-FileUtils.cp  "#{INSTALL_OUT}/setup.exe",  "#{PREFIXD}/#{PKGNAME}"
-FileUtils.mv  "#{STAGE_DIR}/manifest.txt", "#{PREFIXD}/manifest.txt"
-FileUtils.mv  "#{STAGE_DIR}/manifest.xml", "#{PREFIXD}/manifest.xml"
-
+if     File.exist("#{INSTALL_OUT}/setup.exe")
+    FileUtils.cp  "#{INSTALL_OUT}/setup.exe",  "#{PREFIXD}/#{PKGNAME}"
+    FileUtils.mv  "#{STAGE_DIR}/manifest.txt", "#{PREFIXD}/manifest.txt"
+    FileUtils.mv  "#{STAGE_DIR}/manifest.xml", "#{PREFIXD}/manifest.xml"
+else
+    print "\nFAIL:  File does not exist: #{INSTALL_OUT}/setup.exe"
+    exit 99
+end
